+++ CHANGELOG.md
@@ -2,9 +2,10 @@
     - Added Model Orchestra meta estimator
     - Added Delta TF-IDF Transformer
     - Added document frequency smoothing to TF-IDF Transformer
-    - Added Unform neural net weight initializer
+    - Added Uniform neural net weight initializer
     - Fixed missing probabilities in Classification Tree
     - Removed MetaEstimator interface
+    - Added model Wrapper interface
 
 - 0.0.7-alpha
     - Added Support Vector Machine classifier and regressor